idl: Move IDL types from the `anchor-syn` crate to the new IDL crate
