+++ idl/Cargo.toml
@@ -13,18 +13,13 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
-build = [
-    "anyhow",
-    "regex",
-    "serde",
-    "serde_json",
-]
+build = ["anchor-syn", "regex"]
 
 [dependencies]
-anchor-syn = { path = "../lang/syn", version = "0.29.0", features = ["idl-types"] }
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 
 # `build` feature only
-anyhow = { version = "1", optional = true }
+anchor-syn = { path = "../lang/syn", version = "0.29.0", optional = true }
 regex = { version = "1", optional = true }
-serde = { version = "1", features = ["derive"], optional = true }
-serde_json = { version = "1", optional = true }
