+++ idl/src/lib.rs
@@ -4,3 +4,6 @@ pub mod types;
 
 #[cfg(feature = "build")]
 pub mod build;
+
+#[cfg(feature = "build")]
+pub use serde_json;
