+++ lang/Cargo.toml
@@ -33,7 +33,7 @@ idl-build = [
     "anchor-attribute-program/idl-build",
     "anchor-derive-accounts/idl-build",
     "anchor-derive-serde/idl-build",
-    "anchor-syn/idl-build",
+    "anchor-idl/build",
 ]
 init-if-needed = ["anchor-derive-accounts/init-if-needed"]
 interface-instructions = ["anchor-attribute-program/interface-instructions"]
@@ -49,8 +49,8 @@ anchor-derive-accounts = { path = "./derive/accounts", version = "0.29.0" }
 anchor-derive-serde = { path = "./derive/serde", version = "0.29.0" }
 anchor-derive-space = { path = "./derive/space", version = "0.29.0" }
 
-# `anchor-syn` should only be included with `idl-build` feature
-anchor-syn = { path = "./syn", version = "0.29.0", optional = true }
+# `anchor-idl` should only be included with `idl-build` feature
+anchor-idl = { path = "../idl", version = "0.29.0", optional = true }
 
 arrayref = "0.3"
 base64 = "0.21"
