+++ lang/attribute/program/Cargo.toml
@@ -17,7 +17,8 @@ idl-build = ["anchor-syn/idl-build"]
 interface-instructions = ["anchor-syn/interface-instructions"]
 
 [dependencies]
-anchor-syn = { path = "../../syn", version = "0.29.0", features = ["idl-types"] }
+anchor-idl = { path = "../../../idl", version = "0.29.0" }
+anchor-syn = { path = "../../syn", version = "0.29.0" }
 anyhow = "1"
 bs58 = "0.5"
 heck = "0.3"
