+++ lang/attribute/program/src/declare_program/common.rs
@@ -1,4 +1,4 @@
-use anchor_syn::idl::types::{
+use anchor_idl::types::{
     Idl, IdlArrayLen, IdlDefinedFields, IdlField, IdlGenericArg, IdlRepr, IdlSerialization,
     IdlType, IdlTypeDef, IdlTypeDefGeneric, IdlTypeDefTy,
 };
