+++ lang/syn/src/idl/external.rs
